FDS-5 Rename POC module string > text
