--- packages/cascara/src/modules/DataString/index.js
@@ -1,1 +1,0 @@
-export { default } from './DataString';
