+++ packages/cascara/src/modules/DataText/DataText.doc.mdx
@@ -1,0 +1,6 @@
+---
+public: false
+status: 'Draft'
+---
+
+# DataText
