+++ packages/cascara/src/modules/DataText/DataText.js
