+++ packages/cascara/src/modules/DataText/index.js
@@ -1,0 +1,1 @@
+export { default } from './DataText';
